playlist creat, playlist에 add song 기능 구현

+++ src/main/java/com/acorn/melody2/entity/Playlist.java
@@ -2,9 +2,7 @@ package com.acorn.melody2.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
-import lombok.Getter;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Data
@@ -15,10 +13,10 @@ public class Playlist {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "Playlist_ID")
-    private int playlistId;
+    private Long playlistId;
 
     @Column(name = "UserAccount_ID")
-    private int userAccountId;
+    private Long userAccountId;
 
     @Column(name = "Playlist_Name", length = 100)
     private String playlistName;
